Removed shared type formatters in favor of instance variables in CDClassDump2.  This will be better because the type formatters have a delegate, which doesn't work so well when they are shared.  I also want to make the formatting/uniquing of structs more easily testable, so I want it to use a CDClassDump2 object.
This isn't ideal because I've set up the CDOC* methods to have
backpointers to their owners.  This is just a bit ugly to begin with,
but it turns out that CDOCProtocols don't necessarily have a
CDOCSymtab for an owner.  We scan through the protocol section
directly to pick up unreferenced protocols, such as
NSConnectionProtocol and NSProtocolCheckerForwardingMethods.

I think I'll remove the back pointers and just pass the CDClassDump2
object down to the appendToString: methods, and then they can get
the appropriate type formatter from that.