My previous change to CDTypeLexer wasn't necessarily correct, so I switched it back.
Moved the identifier scanning code into a method in an NSScanner
category so that I can use it in -peekIdentifier.  Added that method
so that I can check the first letter (instead of the first character)
for uppercase in the heuristics, which started breaking in one case
now that the members are named.

Added more info (remaining string) when we catch a parsing exception.