Cleaned up type parsing from property attribute string.
Removed a bunch of logs.  Fixed an error with componentsSeparatedByString: used on an empty string,
which returns an array containing the empty string (not what I expected).  Removed TODO item.
Added a debug warning if we find a T attribute that is not at the beginning of the string.

We could still do better.  We parse the type so we can extract the string, and then parse it again
to format it.